app/core/database.py
import psycopg2
from app.core.config import settings

def get_db():
    return psycopg2.connect(
        dbname=settings.DB_NAME,
        user=settings.DB_USER,
        password=settings.DB_PASS,
        host=settings.DB_HOST
    )