requirements.txt
fastapi
uvicorn
psycopg2-binary
sentence-transformers
python-docx
langchain
openai
python-dotenv